webui/requirements.txt
# Core dependencies
fastapi>=0.68.0
uvicorn>=0.15.0
websockets>=10.0
httpx>=0.24.0
pydantic>=2.0.0
python-dotenv>=0.19.0

# LangChain and LLM providers
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-anthropic>=0.1.1
langchain-core>=0.1.22
openai>=1.0.0
anthropic>=0.7.0

# UI and utilities
streamlit>=1.28.0
rich>=13.0.0
tavily-python>=0.5.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.5
pytest-mock>=3.12.0
